# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aura by CannaPost API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brand:
    get:
      operationId: getBrandProfile
      tags: [brand]
      summary: Get the current user's brand profile
      responses:
        "200":
          description: Brand profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
        "404":
          description: Brand profile not found
    post:
      operationId: createBrandProfile
      tags: [brand]
      summary: Create brand profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandProfileInput"
      responses:
        "201":
          description: Created brand profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
    patch:
      operationId: updateBrandProfile
      tags: [brand]
      summary: Update brand profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandProfileUpdate"
      responses:
        "200":
          description: Updated brand profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
...
components:
...
  schemas:
...
    BrandProfile:
      type: object
      required: [id, userId, businessName, industry]
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        industry:
          type: string
        instagramHandle:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        logoBase64:
          type: ["string", "null"]
        watermarkPosition:
          type: ["string", "null"]
        showNameWatermark:
          type: ["boolean", "null"]
        brandColors:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        secondaryColor:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
        backgroundColor:
          type: ["string", "null"]
        fontStyle:
          type: ["string", "null"]
          description: Brand font personality. Legacy rows may contain historical lowercase values.
        brandVoice:
          type: ["string", "null"]
        layoutGrid:
          type: ["string", "null"]
        visualMotif:
          type: ["string", "null"]
        toneOfVoice:
          type: ["string", "null"]
        visualFingerprint:
          type: ["string", "null"]
        targetAudience:
          type: ["string", "null"]
        competitorHandles:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BrandProfileInput:
      type: object
      required: [businessName, industry]
      properties:
        businessName:
          type: string
        industry:
          type: string
        instagramHandle:
          type: string
        location:
          type: string
        timezone:
          type: string
        logoUrl:
          type: string
        logoBase64:
          type: string
        watermarkPosition:
          type: string
        showNameWatermark:
          type: boolean
        brandColors:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        backgroundColor:
          type: string
        fontStyle:
          type: string
          enum: [LUXURY, STREETWEAR, HERITAGE, MODERN, MINIMAL, modern, bold, elegant, casual]
          description: Brand font personality; lowercase values remain accepted for older clients.
        brandVoice:
          type: string
        layoutGrid:
          type: string
          enum: [CENTERED, EDITORIAL, ASYMMETRIC, BOLD]
        visualMotif:
          type: string
          enum: [SMOKE, GEOMETRIC, ORGANIC, CHROME, CRYSTALLINE]
        toneOfVoice:
          type: string
          enum: [PREMIUM, BOLD, EDUCATIONAL, FRIENDLY, SOPHISTICATED]
        visualFingerprint:
          type: string
        targetAudience:
          type: string
        competitorHandles:
          type: string
    BrandProfileUpdate:
      type: object
      properties:
        businessName:
          type: string
        industry:
          type: string
        instagramHandle:
          type: string
        location:
          type: string
        timezone:
          type: string
        logoUrl:
          type: string
        logoBase64:
          type: string
        watermarkPosition:
          type: string
        showNameWatermark:
          type: boolean
        brandColors:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        backgroundColor:
          type: string
        fontStyle:
          type: string
          enum: [LUXURY, STREETWEAR, HERITAGE, MODERN, MINIMAL, modern, bold, elegant, casual]
          description: Brand font personality; lowercase values remain accepted for older clients.
        brandVoice:
          type: string
        layoutGrid:
          type: string
          enum: [CENTERED, EDITORIAL, ASYMMETRIC, BOLD]
        visualMotif:
          type: string
          enum: [SMOKE, GEOMETRIC, ORGANIC, CHROME, CRYSTALLINE]
        toneOfVoice:
          type: string
          enum: [PREMIUM, BOLD, EDUCATIONAL, FRIENDLY, SOPHISTICATED]
        visualFingerprint:
          type: string
        targetAudience:
          type: string
        competitorHandles:
          type: string
...